Upgraded a couple of projects, raised the minimum OS version to 10.7, and removed libstdc++ setting from any projects that needed it

+++ Frameworks/SSEQPlayer/SSEQPlayer/Channel.cpp
@@ -136,7 +136,7 @@ void Channel::UpdatePorta(const Track &trk)
 	{
 		int sq_time = static_cast<uint32_t>(trk.portaTime) * static_cast<uint32_t>(trk.portaTime);
 		long abs_sp = std::abs((long)this->sweepPitch);
-		this->sweepLen = (abs_sp * sq_time) >> 11;
+		this->sweepLen = (uint32_t)((abs_sp * sq_time) >> 11);
 	}
 }
 
@@ -504,7 +504,7 @@ void Channel::Update()
 		modParam = Cnv_Sine(this->modCounter >> 8) * this->modRange * this->modDepth;
 
 		if (!this->modType)
-			modParam = static_cast<int64_t>(modParam * 60) >> 14;
+			modParam = static_cast<int>(static_cast<int64_t>(modParam * 60) >> 14);
 		else
 			// This ugly formula whose exact meaning and workings I cannot figure out is used for volume/pan modulation.
 			modParam = ((modParam & ~0xFC000000) >> 8) | ((((modParam < 0 ? -1 : 0) << 6) | (static_cast<uint32_t>(modParam) >> 26)) << 18);
